adding loop to keep listening for a started livestream
