+++ Chat/ClientApp/src/components/LiveStream.tsx
@@ -1,6 +1,7 @@
 ﻿import React, { useEffect, useState } from 'react';
 import { streamMainStyle } from './styles/Stream.styles';
 import { StreamData } from '../core/actions/LiveStreamActions';
+import AzureMediaPlayer from './AzureMediaPlayer';
 
 import axios from 'axios';
 export interface LiveStreamControlProps {
@@ -15,34 +16,19 @@ const startStream = async (roomId: string): Promise<StreamData> => {
 
 export default (props: LiveStreamControlProps): JSX.Element => {
   const delay = (ms: number) => new Promise((res) => setTimeout(res, ms));
-  const { startLiveStream } = props;
+  const { startLiveStream, liveStreamUrl } = props;
   const [isStreaming, setIsStreaming] = useState(false);
 
   async function getLiveStreamURL() {
-    await delay(10000);
-    let result = await startStream(props.roomId);
-    if (result.liveOutputUrl != null) {
-      startLiveStream(result);
+    while (!isStreaming) {
+      await delay(10000);
+      let result = await startStream(props.roomId);
+      if (result.liveOutputUrl != null) {
+        startLiveStream(result);
+        setIsStreaming(true);
+        break;
+      }
     }
-    let createPlayer = () => {
-      var myOptions = {
-        autoplay: true,
-        controls: true,
-        width: '100%',
-        height: '100%',
-        poster: ''
-      };
-      let _window: any = window;
-      var myPlayer = _window.amp('azuremediaplayer', myOptions);
-
-      return myPlayer;
-    };
-
-    let player = createPlayer();
-
-    return () => {
-      player.dispose();
-    };
   }
 
   useEffect(() => {
@@ -52,20 +38,9 @@ export default (props: LiveStreamControlProps): JSX.Element => {
 
   return (
     <div className={streamMainStyle}>
-      <video
-        id="azuremediaplayer"
-        className="azuremediaplayer amp-default-skin"
-        autoPlay
-        controls
-        width="100%"
-        height="100%"
-        poster="poster.jpg"
-        data-setup='{"nativeControlsForTouch": false}'
-        key={props.liveStreamUrl}
-      >
-        <source src={props.liveStreamUrl} type="application/vnd.ms-sstr+xml" />
-      </video>
-      {props.liveStreamUrl}
+      <div className={streamMainStyle} key={props.liveStreamUrl + props.roomId}>
+        <AzureMediaPlayer liveStreamUrl={props.liveStreamUrl}></AzureMediaPlayer>
+      </div>
     </div>
   );
 };
